fix: dirent's size

+++ include/fs.h
@@ -68,7 +68,7 @@ struct dinode {
 // Directory is a file containing a sequence of dirent structures
 struct dirent {
   uint inum;
-#define DIRSIZE (0x40 - sizeof(ushort))
+#define DIRSIZE (0x40 - sizeof(uint))
   char name[DIRSIZE];
 };
